fix(twig): fix Twig UrlHelpers and split functionnality with new DebugHelpers
