+++ packages/view/src/Charcoal/View/Twig/HelpersInterface.php
@@ -5,11 +5,9 @@ namespace Charcoal\View\Twig;
 /**
  * Defines a collection of Twig helpers.
  *
- * > An array of 'helpers'. Helpers can be global variables or objects, closures
- * > (e.g. for higher order sections), or any other valid Mustache context value.
  * > They will be prepended to the context stack, so they will be available in
- * > any template loaded by this Mustache instance.
- * — {@link https://github.com/bobthecow/mustache.php/wiki#helpers}
+ * > any template loaded by this Twig instance.
+ * — {@link https://twig.symfony.com/doc/3.x/advanced.html#creating-an-extension}
  */
 interface HelpersInterface
 {
